Can I please have my seeding-data be updated with main?

+++ src/main/java/id/ac/ui/cs/advprog/adproggameshop/model/Game.java
@@ -32,7 +32,7 @@ public class Game {
     private User owner;
 
     public Game() {}
-    public Game(String name, float price, String description, int quantity, String category, User owner) {
+    public Game(String name, double price, String description, int quantity, String category, User owner) {
         this.name = name;
         this.price = price;
         this.description = description;
